Fix for heapmemory when it tried to split small memory chunk into two

+++ Engine/Src/SFCore/Util/SFStringCrc64.cpp
@@ -40,7 +40,7 @@ namespace SF
 			context.Radix = 16;
 			context.MaxDigit = 0;
 			context.StringBuffer = strBuffer;
-			context.StringBufferLength = countof(strBuffer);
+			context.StringBufferLength = (int)countof(strBuffer);
 			_IToA(context, (uint32_t)m_Crc);
 			return strBuffer;
 		}
@@ -59,7 +59,7 @@ namespace SF
 			context.Radix = 16;
 			context.MaxDigit = 0;
 			context.StringBuffer = strBuffer;
-			context.StringBufferLength = countof(strBuffer);
+			context.StringBufferLength = (int)countof(strBuffer);
 			_IToA(context, (uint64_t)m_Hash);
 			return strBuffer;
 		}
